refactor: move adding interactions to helper
