+++ views/cypress/tests/item-authoring.spec.js
@@ -18,39 +18,17 @@
 
 import urls from '../utils/urls';
 import selectors from '../utils/selectors';
-import paths from '../utils/paths';
-import { selectUploadLocalAsset } from '../utils/resource-manager';
-import { addShapeToImage } from '../utils/graphic-interactions';
+
+import {
+    addBlockAndInlineInteractions,
+    addCommonInteractions,
+    addMediaInteraction,
+    addGraphicInteractions
+} from '../utils/authoring-add-interactions';
 
 describe('Item Authoring', () => {
     const className = 'Test E2E class';
     const itemName = 'Test E2E item 1';
-    const blockContainer = '_container';
-    const inlineInteractions = {
-        inlineChoiceInteraction: 'inlineChoiceInteraction',
-        textEntryInteraction: 'textEntryInteraction',
-        endAttemptInteraction: 'endAttemptInteraction'
-    };
-    const commonInteractions = {
-        choice: 'choiceInteraction',
-        order: 'orderInteraction',
-        asociate: 'associateInteraction',
-        match: 'matchInteraction',
-        hottext: 'hottextInteraction',
-        gapMatch: 'gapMatchInteraction',
-        slider: 'sliderInteraction',
-        extendedText: 'extendedTextInteraction',
-        upload: 'uploadInteraction'
-    };
-    const mediaInteraction = 'mediaInteraction';
-    const graphicInteractions = {
-        hotspotInteraction: 'hotspotInteraction',
-        graphicOrderInteraction: 'graphicOrderInteraction',
-        graphicAssociateInteraction: 'graphicAssociateInteraction',
-        graphicGapMatchInteraction: 'graphicGapMatchInteraction',
-        selectPointInteraction: 'selectPointInteraction'
-    };
-    const dropSelector = 'div.qti-itemBody.item-editor-drop-area';
     /**
      * Log in
      * Visit the page
@@ -150,53 +128,15 @@ describe('Item Authoring', () => {
             cy.getSettled('.qti-item.item-editor-item.edit-active').should('exist');
             // open inline interactions panel
             cy.get('#sidebar-left-section-inline-interactions').click();
-
-            cy.log('ADDING A-BLOCK INTERACTION');
-            const blockSelector = `[data-qti-class="${blockContainer}"]`;
-            cy.dragAndDrop(blockSelector, dropSelector);
-            // check that widget is initialized
-            cy.getSettled(`${dropSelector} .widget-box.edit-active${blockSelector}`).should('exist');
-            cy.log('A-BLOCK IS ADDED');
-
-            for (const interaction in inlineInteractions) {
-                cy.log('ADDING INTERACTION', interaction);
-                const interactionSelector = `[data-qti-class="${inlineInteractions[interaction]}"]`;
-                cy.dragAndDrop(interactionSelector, dropSelector, [
-                    '.widget-box[data-qti-class="_container"]',
-                    '.widget-box[data-qti-class="_container"] span.qti-word-wrap'
-                ]);
-                // check that widget is initialized (outside item-editor-drop-area)
-                cy.getSettled(`.widget-box.edit-active.widget-${inlineInteractions[interaction]}`).should('exist');
-                cy.getSettled(`.widget-box.edit-active.widget-${inlineInteractions[interaction]} button.widget-ok`).click();
-                cy.log(interaction, 'IS ADDED');
-            }
+            addBlockAndInlineInteractions();
             // close inline interactions panel
             cy.get('#sidebar-left-section-inline-interactions ._accordion').click();
         });
         it('can add common interactions to canvas', () => {
-            for (const interaction in commonInteractions) {
-                cy.log('ADDING INTERACTION', interaction);
-                const interactionSelector = `[data-qti-class="${commonInteractions[interaction]}"]`;
-                cy.dragAndDrop(interactionSelector, dropSelector);
-                // check that widget is initialized
-                cy.getSettled(`${dropSelector} .widget-box.edit-active${interactionSelector}`).should('exist');
-                cy.log(interaction, 'IS ADDED');
-            }
-            // close common interactions panel
-            cy.get('#sidebar-left-section-common-interactions ._accordion').click();
+            addCommonInteractions();
         });
         it('can add media interaction to canvas and upload video', () => {
-            // open common interaction panel
-            cy.get('#sidebar-left-section-common-interactions').click();
-            cy.log('ADDING INTERACTION', mediaInteraction);
-            const interactionSelector = `[data-qti-class="${mediaInteraction}"]`;
-            const videoName = 'sample-mp4-file.mp4';
-            cy.dragAndDrop(interactionSelector, dropSelector);
-            // check that widget is initialized
-            cy.getSettled(`${dropSelector} .widget-box.edit-active${interactionSelector}`).should('exist');
-            // resource selector
-            selectUploadLocalAsset(videoName, `${paths.assetsPath}${videoName}`);
-            cy.log(mediaInteraction, 'IS ADDED');
+            addMediaInteraction();
             // close common interaction panel
             cy.get('#sidebar-left-section-common-interactions ._accordion').click();
         });
@@ -204,34 +144,7 @@ describe('Item Authoring', () => {
         it('can add graphic interactions to canvas and upload image', () => {
             // open graphic interactions panel
             cy.get('#sidebar-left-section-graphic-interactions').click();
-            for (const interaction in graphicInteractions) {
-                cy.log('ADDING INTERACTION', interaction);
-                const interactionSelector = `[data-qti-class="${graphicInteractions[interaction]}"]`;
-                const imageName = 'cats.jpg';
-                const imageOptionName = 'img-option.png';
-                cy.dragAndDrop(interactionSelector, dropSelector);
-                // check that widget is initialized
-                cy.getSettled(`${dropSelector} .widget-box.edit-active${interactionSelector}`).should('exist');
-                selectUploadLocalAsset(imageName, `${paths.assetsPath}${imageName}`);
-
-                // image is loaded to widget
-                cy.getSettled(`.widget-box.edit-active${interactionSelector} .main-image-box image`).should('exist');
-                cy.get(`.widget-box.edit-active${interactionSelector} .image-editor`).then(imageEditor => {
-                    // add shape to image if needed
-                    if (imageEditor.find('li[data-type="rect"]').length) {
-                        addShapeToImage(interactionSelector, 'rect');
-                    }
-                    // add image option if needed
-                    if (imageEditor.find('li.add-option').length) {
-                        cy.get(`.widget-box.edit-active${interactionSelector} .image-editor li.add-option`).click();
-                        selectUploadLocalAsset(imageOptionName, `${paths.assetsPath}${imageOptionName}`);
-                        cy.getSettled(`.widget-box.edit-active${interactionSelector} .source .qti-choice img`).should(
-                            'exist'
-                        );
-                    }
-                });
-                cy.log(interaction, 'IS ADDED');
-            }
+            addGraphicInteractions();
         });
 
         it('can save item with interactions', () => {
